Refactor router and server files to improve type imports and enhance mutator definitions. Remove outdated documentation files to streamline the codebase.
