--- packages/zero-effect/index.ts
@@ -1,4 +1,0 @@
-export * from '@openfaith/zero-effect/effectMutatorBridge'
-export * from '@openfaith/zero-effect/effectMutatorDefs'
-export * from '@openfaith/zero-effect/effectTransaction'
-export * from '@openfaith/zero-effect/zeroLayer'
