+++ packages/zero/mutators.ts
@@ -2,19 +2,14 @@ import { TokenKey } from '@openfaith/adapter-core/layers/tokenManager'
 import type { AuthData, ZSchema } from '@openfaith/zero/zeroSchema.mts'
 import {
   type CustomMutatorEfDefs,
-  convertEffectMutatorsToPromise,
   type EffectTransaction,
-  MutatorAuthError,
-  MutatorValidationError,
-} from '@openfaith/zero-effect'
+  ZeroMutatorAuthError,
+  ZeroMutatorValidationError,
+} from '@openfaith/zero-effect/client'
+import { convertEffectMutatorsToPromise } from '@openfaith/zero-effect/effectMutatorConverter'
 import type { CustomMutatorDefs } from '@rocicorp/zero'
-import { Context, Effect, FiberRefs, type Runtime, RuntimeFlags, Schema } from 'effect'
+import { Effect, Runtime, Schema } from 'effect'
 
-/**
- * App-specific input schemas
- */
-
-// Define the input type for updating a person (basic fields for now)
 export const UpdatePersonInput = Schema.Struct({
   firstName: Schema.String.pipe(Schema.optional),
   id: Schema.String,
@@ -22,9 +17,6 @@ export const UpdatePersonInput = Schema.Struct({
 
 export type UpdatePersonInput = Schema.Schema.Type<typeof UpdatePersonInput>
 
-/**
- * App-specific mutators using Effect-based approach
- */
 export function createMutators(
   authData: Pick<AuthData, 'sub' | 'activeOrganizationId'> | undefined,
 ): CustomMutatorEfDefs<ZSchema, TokenKey> {
@@ -33,26 +25,24 @@ export function createMutators(
       update: (
         tx: EffectTransaction<ZSchema>,
         input: UpdatePersonInput,
-      ): Effect.Effect<void, MutatorAuthError | MutatorValidationError, TokenKey> =>
+      ): Effect.Effect<void, ZeroMutatorAuthError | ZeroMutatorValidationError, TokenKey> =>
         Effect.gen(function* () {
           const tokenKey = yield* TokenKey
 
           console.log(tokenKey)
 
-          // Authentication check
           if (!authData) {
             return yield* Effect.fail(
-              new MutatorAuthError({
+              new ZeroMutatorAuthError({
                 message: 'Not authenticated',
               }),
             )
           }
 
-          // Input validation using Effect Schema
           const validatedInput = yield* Schema.decodeUnknown(UpdatePersonInput)(input).pipe(
             Effect.mapError(
-              (_error): MutatorValidationError =>
-                new MutatorValidationError({
+              (_error): ZeroMutatorValidationError =>
+                new ZeroMutatorValidationError({
                   message: `Invalid input: ${String(_error)}`,
                 }),
             ),
@@ -66,30 +56,21 @@ export function createMutators(
           yield* Effect.log('Person updated successfully', {
             id: validatedInput.id,
           })
-        }) as Effect.Effect<void, MutatorAuthError | MutatorValidationError, TokenKey>,
+        }) as Effect.Effect<void, ZeroMutatorAuthError | ZeroMutatorValidationError, TokenKey>,
     },
   } as const
 }
 
-/**
- * Create Promise-based mutators for client-side use (Zero client expects Promise-based mutators)
- * Provides a mock TokenKey service for client-side use
- */
 export function createClientMutators(
   authData: Pick<AuthData, 'sub' | 'activeOrganizationId'> | undefined,
 ): CustomMutatorDefs<ZSchema> {
   const effectMutators = createMutators(authData)
 
-  // Create a runtime with mock TokenKey service for client-side use
-  const context = Context.make(TokenKey, 'client-token-key')
-  const clientRuntime = {
-    context,
-    fiberRefs: FiberRefs.empty(),
-    runtimeFlags: RuntimeFlags.none,
-  } as Runtime.Runtime<TokenKey>
+  const clientRuntime = Runtime.defaultRuntime.pipe(
+    Runtime.provideService(TokenKey, 'client-token-key'),
+  )
 
   return convertEffectMutatorsToPromise(effectMutators, clientRuntime)
 }
 
-// Export type for the mutators
 export type Mutators = ReturnType<typeof createClientMutators>
